requirements.txt
flake8==6.0.0
Jinja2==3.1.2
markdown-it-py==2.2.0
marko==1.3.0
MarkupSafe==2.1.2
mccabe==0.7.0
mdurl==0.1.2
pycodestyle==2.10.0
pyflakes==3.0.1
Pygments==2.14.0
python-frontmatter==1.0.0
PyYAML==5.1
rich==13.3.3
